Query for both static and dynamic sample rate, trust dynamic
The SACD input plugin reports the native sample rate of DSD data files
as the "samplerate" info field, which requires us to query the dynamic
information unconditionally for the decoded sample rate.

This is now done for all forms of input and the dynamic sample rate is
preferred if present.