Handle device resets for Direct2D
In both the image caching thread and the main rendering thread, uses of
Direct2D needs to be tested for losing the device on EndDraw, causing
all future renderings to fail until the render target and all associated
resources are recreated.

Additionally, we change how the waveform bitmap is created, removing the
need for the image cache renderer to know about the render target of the
frontend renderer.