Expose peak information to effect
The shader semantics "CHANNELMAGNITUDE" and "TRACKMAGNITUDE" are
float4:s representing the amplitude of the current channel and the whole
track.

R is the minimum amplitude, G is the maximum amplitude, B is the highest
RMS value, A is 1.0 if the data is present.

This can be used to rescale the vertical texcoord to display a scaled
and displaced range, ensuring that the edges of the display are
peak-to-peak.

In most effects, add something akin to the following at the top next to
all the other semantics:
float4 chan_mag : CHANNELMAGNITUDE;
float4 track_mag : TRACKMAGNITUDE;

Add something akin to the following in the main PS function:
input.tc.y = (input.tc.y + 1)/2 * (chan_mag.g - chan_mag.r) + chan_mag.r;

This will map the range [-1,1] to [min_peak,max_peak].

Other effects are possible, like a symmetric scale by the largest
absolute value:
input.tc.y = input.tc.y * max(abs(chan_mag.r), abs(chan_mag.g));

This will map the range [-1,1] to [-largest_peak,largest_peak].