Report incremental scan progress
The scanning functionality will now prepare intermediary waveforms and
periodically report those back to the seekbar, including information on
how many buckets of the waveform are valid. It will fill invalid buckets
with zeroes to avoid disturbing displays.

This behavior is off-by-default and can be enabled through an advanced
setting labelled "Incremental update of waveforms being scanned".

Note that this might have a negative impact on scanning performance when
enabled, as it does a fair bit of work in preparing these partial
results. Caveat enabler.