Restore lost LFE channel from stereo/mono mixdown
LFE mixdown factor is now 0.5 per channel.
Removing the LFE channel completely from mixdowns is a bit of a lie, as
the content, even while purely supplemental, has some meaning in the
context of a track.

As every mixdown is a reinterpretation of the data, having an effect
over-represented is better than completely omitting it, particularly as
there is now shader parameters to adjust the displayed range of a
waveform.

ITU recommendation BS.775 specifies a choice of four mixin factors for
LFE content: 0.7071, 0.5, 0.0 and 'reserved', to be specified by the
broadcaster. As any such information is not available, I've selected 0.5
as the coefficient due to it being the lowest non-zero choice, avoiding
over-attenuation while still representing the content somehow.