Protect against corrupt tracks that open
Some corrupt files (Matroska with chapters) contain enough information
to deduce sample rate and evade all other early detection of failure and
get as far as attempting to decode the first chunk.

The first chunk is special in that it's the one used by the component to
determine the channel count, and if that's not set, an uninitialized
value is used for the number of channels in the silence chunks emitted.

This failure is now detected and a mono silence waveform is emitted.